--- scripts/__init__.py
@@ -1,6 +1,0 @@
-__author__ = "The Clemente Lab"
-__copyright__ = "Copyright (c) 2018 The Clemente Lab"
-__credits__ = ['Alexander Kyimpopkin', 'David Wallach', 'Jose C. Clemente']
-__license__ = "GPL"
-__maintainer__ = "Alexander Kyimpopkin"
-__email__ = "alexanderkyim@gmail.com"
